refactor: moved models inside server
