+++ src/app/api/users/get-users/route.ts
@@ -2,7 +2,7 @@ import { Databases } from 'appwrite'
 import { createInstace } from '@/client/services/appwrite'
 import env from '@/configs/environment'
 import { NextResponse } from 'next/server'
-import { User } from '@/model/user'
+import { User } from '@/server/model/user'
 const client = createInstace()
 const databases = new Databases(client)
 const collectionId = env.appwriteCollectionId.user
